requirements.txt
Flask>=0.10.1
kafka-python==0.9.3
gevent
avro==1.7.7
